edit: Admin 데이터 관련 함수 api.ts에 분리
